A program that finds and prints the sum of the even-valued terms, followed by a new line. starting from 0 and 1.

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -1,16 +1,17 @@
 #include <stdio.h>
 
 /**
- * main - finds and prints the first 98 Fibonacci numbers,
- * starting with 1 and 2
+ * main -  finds and prints the sum of the even-valued terms
+ * starting with 0 and 1
  * followed by a new line
+ *
  * Return: ALways 0 (Success)
  */
 
 int main(void)
 {
 	unsigned long fibo1 = 0, fibo2 = 1, sum;
-	long int total;
+	float total;
 
 	while (1)
 	{
@@ -29,7 +30,7 @@ int main(void)
 		fibo1 = fibo2;
 		fibo2 = sum;
 	}
-	printf("%ld\n", total);
+	printf("%.0f\n", total);
 
 	return (0);
 }
